chore: Move jnigen test to gradle plugin tests
