+++ compiling/gdx-jnigen/build.gradle
@@ -1,7 +1,6 @@
 dependencies {
     implementation "com.github.javaparser:javaparser-core:3.14.14"
     api project(':gdx-jnigen-commons')
-    testImplementation "junit:junit:4.12"
 }
 
 eclipse {
